requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
click==7.1.2
Django==3.1.5
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.2
flake8==3.8.4
furl==2.1.0
isort==5.7.0
mccabe==0.6.1
mypy-extensions==0.4.3
orderedmultidict==1.0.1
pathspec==0.8.1
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.5
regex==2020.11.13
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.5.4
typing-extensions==3.7.4.3
